+++ silo-core/test/foundry/utils/liquidationHelper/LiquidationHelper_1token.i.sol
@@ -36,9 +36,9 @@ contract LiquidationHelper1TokenTest is LiquidationHelperCommon {
     forge test --ffi --mt test_executeLiquidation_1_token -vvv
     */
     function test_executeLiquidation_1_token(
-        uint32 _addTimestamp
+        uint64 _addTimestamp
     ) public {
-        vm.assume(block.timestamp + _addTimestamp < type(uint32).max);
+        vm.assume(block.timestamp + _addTimestamp < type(uint64).max);
         vm.warp(block.timestamp + _addTimestamp);
 
         (uint256 collateralToLiquidate, uint256 debtToRepay,) = partialLiquidation.maxLiquidation(BORROWER);
