+++ silo-core/test/foundry/utils/liquidationHelper/ManualLiquidationHelper_2tokens.i.sol
@@ -31,8 +31,8 @@ contract ManualLiquidationHelper2TokensTest is ManualLiquidationHelperCommon {
     /*
     FOUNDRY_PROFILE=core-test forge test --ffi --mt test_executeLiquidation_2_tokens -vvv
     */
-    function test_executeLiquidation_2_tokens(uint32 _addTimestamp) public {
-        vm.assume(block.timestamp + _addTimestamp < type(uint32).max);
+    function test_executeLiquidation_2_tokens(uint64 _addTimestamp) public {
+        vm.assume(block.timestamp + _addTimestamp < type(uint64).max);
         vm.warp(block.timestamp + _addTimestamp);
 
         (uint256 collateralToLiquidate, uint256 debtToRepay,) = partialLiquidation.maxLiquidation(BORROWER);
